@@ -1,106 +1,77 @@
 {
-	"name": "logigator-editor",
-	"version": "1.1.2",
-	"description": "Logigator Editor",
-	"license": "MIT",
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/logigator/logigator.git"
-	},
-	"scripts": {
-		"start": "yarn run start:web",
-		"start:web": "yarn run build-custom-build-scripts && yarn run cp-sim-wasm-assets && env-cmd -e dev-web --use-shell \"ng serve -c dev-web\"",
-		"start:electron": "yarn run build-custom-build-scripts && env-cmd -e dev-electron --use-shell \"electron-rebuild && npm-run-all -p ng:serve:electron electron:serve\"",
-		"build": "yarn run build:web",
-		"build:web": "yarn run build-custom-build-scripts && yarn run cp-sim-wasm-assets && env-cmd -e prod-web --use-shell \"ng build -c prod-web \" && yarn run wasm-js-cache-bust",
-		"build:electron:win": "yarn run prepackage:electron && yarn run package:electron:win",
-		"build:electron:linux": "yarn run prepackage:electron && yarn run package:electron:linux",
-		"build:electron:win64": "yarn run prepackage:electron && node custom-build-scripts/dist/package-app.js win32 x64",
-		"build:electron:win32": "yarn run prepackage:electron && node custom-build-scripts/dist/package-app.js win32 ia32",
-		"build:electron:win-arm64": "yarn run prepackage:electron && node custom-build-scripts/dist/package-app.js win32 arm64",
-		"build:electron:linux64": "yarn run prepackage:electron && node custom-build-scripts/dist/package-app.js linux x64",
-		"build:electron:linux32": "yarn run prepackage:electron && node custom-build-scripts/dist/package-app.js linux ia32",
-		"build:electron:linux-arm64": "yarn run prepackage:electron && node custom-build-scripts/dist/package-app.js linux arm64",
-		"prepackage:electron": "yarn run build-custom-build-scripts && env-cmd -e prod-electron --use-shell \"ng build -c prod-electron && yarn run electron:tsc && node custom-build-scripts/dist/create-dist-package.js && yarn run electron-npm-i\"",
-		"package:electron:win": "node custom-build-scripts/dist/package-app.js win32 x64 && node custom-build-scripts/dist/package-app.js win32 ia32 && node custom-build-scripts/dist/package-app.js win32 arm64",
-		"package:electron:linux": "node custom-build-scripts/dist/package-app.js linux x64 && node custom-build-scripts/dist/package-app.js linux ia32 && node custom-build-scripts/dist/package-app.js linux arm64",
-		"cp-sim-wasm-assets": "node custom-build-scripts/dist/copy-simulation-wasm-assets.js",
-		"wasm-js-cache-bust": "node custom-build-scripts/dist/wasm-js-cache-bust.js",
-		"electron-npm-i": "cd dist && yarn install --silent",
-		"ng:serve:electron": "ng serve -c dev-electron",
-		"electron:serve": "wait-on http-get://localhost:8202/ && yarn run electron:tsc && electron . --inspect=5858",
-		"electron:debug": "yarn run build-custom-build-scripts && env-cmd -e dev-electron --use-shell \"electron-rebuild --debug --force && yarn run ng:serve:electron\"",
-		"electron:tsc": "tsc -p tsconfig.electron.json",
-		"build-custom-build-scripts": "tsc -p custom-build-scripts/tsconfig-custom-build-scripts.json",
-		"lint": "ng lint"
-	},
-	"main": "./dist/electron/main.js",
-	"private": true,
-	"dependencies": {
-		"@angular/animations": "^11.2.14",
-		"@angular/common": "^11.2.14",
-		"@angular/compiler": "^11.2.14",
-		"@angular/core": "^11.2.14",
-		"@angular/forms": "^11.2.14",
-		"@angular/platform-browser": "^11.2.14",
-		"@angular/platform-browser-dynamic": "^11.2.14",
-		"@angular/router": "^11.2.14",
-		"@logigator/logigator-simulation": "^1.2.0",
-		"@ngx-translate/core": "^13.0.0",
-		"bootstrap": "^4.6.1",
-		"form-data": "^4.0.0",
-		"github-markdown-css": "^4.0.0",
-		"ngx-toastr": "^13.2.1",
-		"node-fetch": "^2.6.6",
-		"pixi.js": "^5.3.9",
-		"run-script-os": "^1.1.6",
-		"rxjs": "~6.5.4",
-		"tslib": "^2.0.0",
-		"uuid": "^8.3.2",
-		"zone.js": "~0.10.2"
-	},
-	"devDependencies": {
-		"@angular-builders/custom-webpack": "^11.1.1",
-		"@angular-devkit/build-angular": "~0.1102.15",
-		"@angular/cli": "^11.2.15",
-		"@angular/compiler-cli": "^11.2.14",
-		"@angular/language-service": "^11.2.14",
-		"@ngtools/webpack": "^11.2.15",
-		"@types/jasmine": "~3.6.0",
-		"@types/jasminewd2": "^2.0.8",
-		"@types/node": "^16.11.10",
-		"@types/node-fetch": "^2.5.12",
-		"codelyzer": "^6.0.0",
-		"electron": "^16.0.2",
-		"electron-packager": "^15.4.0",
-		"electron-rebuild": "^3.2.5",
-		"electron-reload": "^1.5.0",
-		"env-cmd": "^10.1.0",
-		"jasmine-core": "~3.6.0",
-		"jasmine-spec-reporter": "~5.0.0",
-		"karma": "~6.3.9",
-		"karma-chrome-launcher": "~3.1.0",
-		"karma-coverage-istanbul-reporter": "~3.0.2",
-		"karma-jasmine": "~4.0.0",
-		"karma-jasmine-html-reporter": "^1.5.0",
-		"markdown-loader": "^6.0.0",
-		"npm-run-all": "^4.1.5",
-		"protractor": "~7.0.0",
-		"raw-loader": "^4.0.2",
-		"tslint": "~6.1.3",
-		"typescript": "4.0.2",
-		"wait-on": "^6.0.0",
-		"webpack-bundle-analyzer": "^4.5.0",
-		"@logigator/webpack-preprocessor-loader": "^1.1.4"
-	},
-	"electronProdDependencies": {
-		"form-data": "^4.0.0",
-		"node-fetch": "^2.6.6",
-		"tslib": "^2.2.0",
-		"@logigator/logigator-simulation": "^1.2.0"
-	},
-	"developmentInfo": {
-		"": "For Development you should use the local version of shared-cops to use live reload from the libary",
-		"@logigator/logigator-simulation": "file:../../logigator-simulation"
-	}
+  "name": "logigator-editor",
+  "version": "1.1.2",
+  "description": "Logigator Editor",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/logigator/logigator.git"
+  },
+  "scripts": {
+    "start": "yarn run start:web",
+    "start:web": "yarn run build-custom-build-scripts && yarn run cp-sim-wasm-assets && env-cmd -e dev-web --use-shell \"ng serve -c dev-web\"",
+    "build": "yarn run build:web",
+    "build:web": "yarn run build-custom-build-scripts && yarn run cp-sim-wasm-assets && env-cmd -e prod-web --use-shell \"ng build -c prod-web \" && yarn run wasm-js-cache-bust",
+    "cp-sim-wasm-assets": "node custom-build-scripts/dist/copy-simulation-wasm-assets.js",
+    "wasm-js-cache-bust": "node custom-build-scripts/dist/wasm-js-cache-bust.js",
+    "build-custom-build-scripts": "tsc -p custom-build-scripts/tsconfig-custom-build-scripts.json",
+    "lint": "ng lint"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^17.0.8",
+    "@angular/common": "^17.0.8",
+    "@angular/compiler": "^17.0.8",
+    "@angular/core": "^17.0.8",
+    "@angular/forms": "^17.0.8",
+    "@angular/platform-browser": "^17.0.8",
+    "@angular/platform-browser-dynamic": "^17.0.8",
+    "@angular/router": "^17.0.8",
+    "@logigator/logigator-simulation": "^1.2.0",
+    "@ngx-translate/core": "^15.0.0",
+    "bootstrap": "^4.6.2",
+    "form-data": "^4.0.0",
+    "github-markdown-css": "^4.0.0",
+    "ngx-toastr": "^18.0.0",
+    "node-fetch": "^3.3.2",
+    "pixi.js": "^5.3.9",
+    "run-script-os": "^1.1.6",
+    "rxjs": "^7.8.1",
+    "tslib": "^2.6.2",
+    "uuid": "^9.0.1",
+    "zone.js": "~0.14.2"
+  },
+  "devDependencies": {
+    "@angular-builders/custom-webpack": "^17.0.0",
+    "@angular-devkit/build-angular": "^17.0.8",
+    "@angular/cli": "^17.0.8",
+    "@angular/compiler-cli": "^17.0.8",
+    "@angular/language-service": "^17.0.8",
+    "@ngtools/webpack": "^17.0.8",
+    "@types/jasmine": "^5.1.4",
+    "@types/jasminewd2": "^2.0.13",
+    "@types/node": "^20.10.5",
+    "@types/node-fetch": "^2.6.10",
+    "codelyzer": "^6.0.2",
+    "env-cmd": "^10.1.0",
+    "jasmine-core": "~3.99.0",
+    "jasmine-spec-reporter": "~5.0.2",
+    "karma": "~6.4.2",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~4.0.2",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "markdown-loader": "^8.0.0",
+    "npm-run-all": "^4.1.5",
+    "protractor": "~7.0.0",
+    "tslint": "^6.1.3",
+    "typescript": "5.2.2",
+    "wait-on": "^6.0.1",
+    "webpack": "^5.59.1",
+    "webpack-bundle-analyzer": "^4.10.1"
+  },
+  "developmentInfo": {
+    "": "For Development you should use the local version of shared-cops to use live reload from the libary",
+    "@logigator/logigator-simulation": "file:../../logigator-simulation"
+  }
 }
