@@ -5,12 +5,9 @@ import {ReloadQuestionComponent} from '../../popup-contents/reload-question/relo
 import {TranslateService} from '@ngx-translate/core';
 import {UserService} from '../../../services/user/user.service';
 import {ProjectsService} from '../../../services/projects/projects.service';
-import {Observable} from 'rxjs';
-import {ElectronUpdateService} from '../../../services/electron-update/electron-update.service';
 import {PopupService} from '../../../services/popup/popup.service';
 import {environment} from '../../../../environments/environment';
 import {SimulationManagementService} from '../../../services/simulation/simulation-management/simulation-management.service';
-import {ElectronService} from '../../../services/electron/electron.service';
 
 @Component({
 	selector: 'app-settings-dropdown',
@@ -26,24 +23,17 @@ export class SettingsDropdownComponent implements OnInit {
 	public showDropDown = true;
 	public currentLang: string;
 
-	public isNewVersionAvailable$: Observable<boolean>;
-
 	constructor(
 		public theming: ThemingService,
 		public simulationMgmt: SimulationManagementService,
 		private popupService: PopupService,
 		private translation: TranslateService,
 		private user: UserService,
-		private projects: ProjectsService,
-		@Optional() private electronUpdateService: ElectronUpdateService,
-		@Optional() private electronService: ElectronService
+		private projects: ProjectsService
 	) {}
 
 	ngOnInit(): void {
 		this.currentLang = this.translation.currentLang;
-
-		// #!electron
-		this.isNewVersionAvailable$ = this.electronUpdateService.isNewVersionAvailable$;
 	}
 
 	public get isLoggedIn(): boolean {
@@ -65,32 +55,12 @@ export class SettingsDropdownComponent implements OnInit {
 	}
 
 	public accountSettings() {
-		// #!web
 		window.open(`${environment.homeUrl}/my/account/profile`, '_blank');
-		// #!electron
-		this.electronService.shell.openExternal(`${environment.homeUrl}/my/account/profile`);
 	}
 
 	public privacyPolicy() {
-		// #!web
 		window.open(`${environment.homeUrl}/privacy-policy`, '_blank');
-		// #!electron
-		this.electronService.shell.openExternal(`${environment.homeUrl}/privacy-policy`);
 	}
-
-	// #! ELECTRON === 'true'
-	public async logout() {
-		if (await this.projects.askToSave()) {
-			await this.projects.newProject();
-			await this.user.logout();
-		}
-		this.close();
-	}
-
-	updateClick() {
-		this.electronService.shell.openExternal(`${environment.homeUrl}/download`);
-	}
-	// #!endif
 
 	public async showReloadPopup() {
 		this.showDropDown = false;
